requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
numpy
scipy
git+https://github.com/suno-ai/bark.git
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers
pytest
nltk
librosa
scikit-learn
soundfile
gtts
requests
gradio